add sqs.delete.enabled for source connector

+++ src/main/java/com/nordstrom/kafka/connect/sqs/SqsSourceConnectorTask.java
@@ -126,7 +126,11 @@ public class SqsSourceConnectorTask extends SourceTask {
     final String receipt = record.sourceOffset().get( SqsConnectorConfigKeys.SQS_MESSAGE_RECEIPT_HANDLE.getValue() )
         .toString() ;
     log.debug( ".commit-record:url={}, receipt-handle={}", config.getQueueUrl(), receipt ) ;
-    client.delete( config.getQueueUrl(), receipt ) ;
+
+    if(config.getDeleteEnabled()){
+      client.delete( config.getQueueUrl(), receipt ) ;
+    }
+
     super.commitRecord( record ) ;
   }
